Ensure resulting /usr/local is not user writable
Because we should do as little modification to the system as possible.

Even though /usr/local doesn't exist on a virgin OS X, many tools create it, and some of them expect /usr/local to be root owned for security reasons.

This modification fixes Airfoil on systems where it is installed.

The result of this is we can no longer create new directories in the HOMEBREW_PREFIX unless we check and prompt for sudo first when required.