Added a Cask for CocoaDialog
Fixes #2499

CocoaDialog is an OS X application that allows the use of common GUI
controls such as file selectors, text input, progress bars, yes/no
confirmations and more with a command-line application. It requires no
knowledge of Cocoa, and is ideal for use in shell and Perl scripts
(or Ruby, or Python, or... etc).

See also http://mstratman.github.io/cocoadialog/.

This PR introduces a Cask for CocoaDialog from version 2.1.1 on
(current upcoming version 3.0.0 is still beta).