remove homebrew-fork Version class
replace with Gem::Version for OS X release comparisons.

Later we will need to pull in something like the Version class or
build something similar, to support a greater understanding of the
value of the version stanza (which is not solely numeric, as
Gem::Version expects).  For now, we don't need the complexity of
this class, as we aren't using it.