Add Lingon X 2.0, moved Lingon X 1.0 to versions
Per the discussion on caskroom/homebrew-cask#8844, Lingon X v1.0 will
be moved to caskroom/homebrew-versions and v2.0 will replace v1.0 here

(+2 squashed commits)

[280f122] Add OS version conditional to Lingon X Cask
Added a conditional to determine if we should install Lingon 1 or Lingon
2 depending on the user's version of OS X.

[80d3909] Add Lingon X.app v2.0
Lingon X version 1.0 already exists as lingon-x but Lingon X 2.0 is an
entirely new version of Lingon X which only runs on Yosemite 10.10 or
higher. In my opinion, Lingon X 1.0 is still a valid Cask for this very
reason, and I think including 2.0 as lingon-x2 is the best way to handle
this updated version.