Downgrade Bartender to version 1.2.39
This commit changes the version and sha256 stanzas.

I updated this cask to 1.2.40 after following a prompt in the app. I
did not realise I was accepting testing versions of the app, and
version 1.2.40 was not a normal stable release. Sorry to anyone who
has been negatively affected by running the testing version in this
period.