UniFi Controller - Upgrading to v4.7.5
Updating the unifi-controller cask to the latest version (4.7.5).

This also updates the homepage URL. The previous URL currently links to
the main Ubiquiti downloads page. The updated URL links specifically to
the UniFi Controller downloads page.

A small change was made to the spelling of UniFi, which was incorrectly
spelled with a lowercase f. This is purely a stylistic change as the
previous spelling was not causing any problems.