Add iMouseTrick v1.1 (#40161)
* Add iMouseTrick v1.1

iMouseTrick is a program designed to perform only one function: click
many times and as fast as you want. When this program is useful? For
example, when you are playing a game where you have to do a lot of
clicks and you're not fast enough or you don't want to lose your finger
clicking all the time. The limit is up to you.

* Update imousetrick.rb

* Change `zap quit` to `uninstall quit`.

* Update imousetrick.rb