Fix Factor 0.97 path environment variable caveat (#44574)
As Factor is now installed as a suite, its staged path will be empty and
cannot be used in an environment variable. Instead, users should be
directed to add the installed application directory to their PATH in
order to use the "factor" binary command.

Note that we cannot add a "binary" stanza for "factor" as it does not
work as a symlink and requires its surrounding files to be in the same
directory.