Delete bitlord (#56325)
* Delete bitlord

as per discussion in https://github.com/Homebrew/homebrew-cask/issues/56223

"this is exactly the same situation as with filezilla a few days ago. there is a clean version, and an 'evil' adware installer version. we link to the clean version, which is good. however, the clean version always lags behind by one version, and if you go to their website and click the big fat green download button, you get the adware installer. (i believe its the very same adware as filezilla). needless to say, i think this cask should be removed, or at the very least an explanatory warning caveat should be added."

* Delete bitlord.rb