Delete icecat.rb (#64333)
i am proposing to delete this. this hasn't been supported for years on macOS via binary releases, as written on the homepage. and using a browser that hasn't been updated for years is probably a super stupid idea. i am thinking since our policy to remove software that has e.g. moved into the appstore and is only available for download in outdated version, the same rule applies here. please double check.