Updated inkscape.rb to remove X11 dependency (#69345)
* Update inkscape.rb to remove X11 dependency

Updated the inkscape dmg to 1.0 beta 1 as it is native Mac OS X with updated GTK+.

This is the first Mac OS native inkscapre application.

* Updated cask to delete files created by installer

Delete files created by installer.

* Update inkscape.rb

* Update inkscape.rb

* Update inkscape.rb

* Updated scribus to current version 1.5.5

The existing version 1.4.8 of scribus is very old and has bee updated for sometime. It needs to be updated.

* Removed changes from scribus.rb

* Change SHA hash and download link

Updated to latest download link and SHA-265 hash as CI was giving error.

* Update inkscape.rb

* Update inkscape.rb

* Update inkscape.rb

* Update inkscape.rb