Delete avira-vpn.rb (#72726)
this is a very old version, so according to our guidelines it should be removed.

it will install version 2.8.2 which is about 18 months old.

the newest version is 2.18.1. 
the newest version is available on the Mac App Store and e.g. as part of the 'avira free security suite'. 

however, i wasn't able to find any standalone download to update the cask. please double check.