Use vendor provided shim for p4merge (#82649)
P4Merge comes with a vendor provided shim that interacts with the main P4Merge
application and that terminates when the window is closed. This helps e.g.
merging several files in a row when P4Merge is launched from a SCM tool like
Git

See https://community.perforce.com/s/article/2848