rustrover 2024.1,241.15989.199
RustRover is now released but the existing `livecheck` block restricts
the check to EAP (unstable) versions and incorrectly returns
2024.1,241.15989.180 (the newest EAP release) as the latest version.
This updates the `livecheck` block to only check stable (`release`)
versions going forward and updates the cask to 2024.1,241.15989.199.

The stable dmg files use the version (`2024.1`) instead of the build
(`241.15989.199`), so we unfortunately have to use `sha256 :no_check`
now, as the dmg can be updated to a new build without any change
to the file name.