fl-studio: update url, livecheck
The cask URL for `fl-studio` doesn't work locally or on CI, so this
updates it to the current URL from the redirecting macOS link on the
download page. Even with this change, it was still necessary to use
a `:browser` user agent and add a referer to get this to work.

The existing `livecheck` block for `fl-studio` returns an `Unable to
get versions` error, as the request is blocked by Cloudflare. As with
the dmg URL, we have to set a `:browser` user agent and referer for
this to work. We can't set the referer in livecheck yet, so we have
to work around this for now.

This addresses the issue by updating the `livecheck` block to parse
the version from the URL that the download page uses to fetch
version information. Unfortunately the server returns JavaScript, so
we have to extract the JSON data from the content, unescape the
string, and parse it. The data only contains FL Studio releases at
the moment but I added some safeguards to help ensure we only return
appropriate versions if the data changes in the future.