quakespasm: add livecheck
livecheck uses the `Sourceforge` strategy to identify version
information for `quakespasm` and this currently works fine but there
are various subdirectories in the SourceForge project and we're only
interested in the `Mac` directory. This adds a `livecheck` block with
a `url` that uses the `path` query string parameter to restrict the
RSS feed to the `Mac` subdirectory, so we don't have to worry about
other files pushing the Mac files out of the RSS feed.