font-gilbert: update livecheck
The existing `livecheck` block for `font-gilbert` is returning an
`Unable to get versions` error, as the GitHub releases HTML no longer
contains asset links but this hasn't been updated to use the
`GithubLatest` or `GithubReleases` strategy.

This updates the `livecheck` block to use the `GithubReleases`
strategy and matches the version information from asset filenames.
It's currently necessary to use `GithubReleases`, as all releases are
marked as "pre-release" (i.e., there's no "latest" release). We
should be able to switch to the `GithubLatest` strategy if/when
there's a "latest" release in the future.