garagesale: update livecheck
The `#VersionHistory` fragment in the `livecheck` block URL doesn't
impact the page content and the regex is matching a dmg URL from a
different part of the page, so this removes the URL fragment for the
sake of simplicity. This also tweaks the regex to bring it more in
line with typical regex patterns (and so it will continue to match
if the URL doesn't contain a forward slash in the future).