spotify 1.2.45.454
Besides updating `spotify` to the latest version, this updates the
`livecheck` block to remove the `strategy` block. The regex in the
existing `strategy` block no longer matches, as the version format
is now simply `1.2.45.454`. The `ExtractPlist` strategy will
automatically use the plist version, so a `strategy` block shouldn't
be necessary now.