openmsx: update livecheck
This reworks the `strategy` block logic to use the typical pattern
for working with Git tags when the regex only uses one capture group.
I've also pared down the version delimiter character class to only
`[._]`, as the `RELEASE` tags don't use a hyphen delimiter and we're
not handling a hyphen in the `strategy` block's `#tr` argument. We
could update the `#tr` argument to also replace hyphens but the
aforementioned approach should be fine for now (and it's in line with
similar `livecheck` blocks).