ci_matrix: Rework depends_on macos array handling
CodeQL complains about parts of the `DEPENDS_ON_MACOS_ARRAY_MEMBER`
regex snippet: `This part of the regular expression may cause
exponential backtracking on strings starting with
'depends_on macos: [:' and containing many repetitions of '!:'.` This
error occurs in relation to the repeating `[^\\s",]` and trailing
`\s` parts of this regex.

I'm not very familiar with the issue CodeQL is flagging but this
reworks handling of `depends_on macos:` arrays in such a way that
CodeQL doesn't return any errors. In the process of addressing this,
I updated the array member regex so it's also capable of matching
strings with a leading comparator (e.g., `">= :big_sur"`) instead of
only symbols like `:big_sur`.

I also opted to simplify the case condition regex and inline the array
member regex, as it may be easier to understand compared to having a
separate regex constant.