post-haste: update livecheck
The existing `livecheck` block for `post-haste` naively inserts dots
into a dotless version string. This approach only works as expected
if all the version parts are a single digit (i.e., a version like
`1.23.4` would be incorrectly parsed as `1.2.3`). It may be that
upstream is careful to only use versions where each numeric part is a
single digit but this can't be maintained forever, as there are a
finite number of available versions before having to add another digit
(e.g., a 10+ major version, more than four numeric parts, etc.).

This avoids the issue by matching versions with dots from the
"Version History" section of the Post Haste support page. This
provides both the dotted and dotless version (which has more numbers),
so we don't have to make assumptions about the difference between the
two.