aerial@beta: update livecheck
The existing `livecheck` block for `aerial@beta` checks the releases
page, searching for only pre-release releases and matching beta
versions. This is an older `livecheck` block that hasn't been updated
to use `GithubReleases`, so this updates it accordingly.

This also makes the beta suffix optional, so this will surface stable
versions in addition to beta versions, like some other casks for
unstable versions.