bettertouchtool 4.997,2024122102
The second part of the current version in the `bettertouchtool` cask
has four digits instead of the usual three, so newer versions are
being treated as lower than the cask version and autobump has been
failing to update this cask. This updates the cask version to
4.997,2024122102, which is the newest version (the `livecheck` block
returns the correct latest version).