squidman: update livecheck, zap
This updates the `livecheck` block for `squidman` to check the
homepage, which contains download links. This is a little closer to
the cask `url` source than the release notes. We can technically
check the /resources/downloads directory listing page but there may
be something to be said for checking the homepage downloads, as a dmg
file may be uploaded before it's considered released.

This also updates the `zap` to account for related app files, as well
as `squid` files that newer versions of Squidman will remove using
"Uninstall Squid" from the main menu.