transmit: downgrade to 5.10.6, update livecheck
The existing `livecheck` block for `transmit` is returning 5.10.7 as
the latest version, as the Release Integrity page contains information
for 5.10.7. However, the 5.10.7 link is currently giving a 404
response and the redirecting "latest" link on the main download page
points to 5.10.6 instead. The cask was recently updated to 5.10.7 but
upgrades are failing because the cask `url` isn't accessible.

This downgrades the cask to 5.10.6 until 5.10.7 is available and
updates the `livecheck` block to check the "latest" link instead.