heroic: add livecheck
There appears to be a large enough gap between when a `heroic` version
is tagged and the GitHub release is created to cause a problem for the
autobump workflow. This adds a `GithubLatest` `livecheck` block to
ensure that autobump won't try to update to a new version until the
corresponding release is available.