1password@beta: update livecheck
The existing `livecheck` block for `1password@beta` is returning an
`Unable to get versions` error, as the content on the page must have
changed in a way that the regex no longer matches anything. This
updates the regex to match the beta versions as they currently
appear on the page.