ontime 3.14.4
This updates `ontime` to the newest version, 3.14.4, and removes the
rogue dot from the tag in the URL, since upstream is using the
expected `v1.2.3` tag format again (after the one-off typo in
`v.3.14.3`). `ontime` is in the autobump list but it's failing to
update because we have to manually update the URL, so this should get
it back on track.