airfoil: update livecheck
The existing `livecheck` block for `airfoil` uses `MacOS.full_version`
in the `url`, so we don't have to maintain a `livecheck` block in
each on_system block. However, `MacOS` isn't something that
`SimulateSystem` handles, so it will always be the macOS version from
the execution environment and that can cause unexpected behavior.
`MacOS` also doesn't exist on Linux, so this causes an error when the
`livecheck` block isn't guarded behind a macOS on_system block.

This moves the existing `livecheck` block into each on_system block
and replaces `MacOS.full_version` with a hardcoded macOS version. I've
used the highest supported macOS version for the ACE release and 999
as an arbitrarily high version for the ARK release, as we're already
doing in the `farrago` and `fission` casks.