cernbox: downgrade to 5.2.1.13464
This downgrades `cernbox` to the latest stable version, 5.2.1.13464.
5.3.2.15486 is a beta version, so we shouldn't have updated the cask
to 5.3.x versions yet. The filename format is the same for both
stable and unstable versions, so livecheck is returning a beta
version as the newest version because it's highest.

With that in mind, this updates the `livecheck` block to check the
first-party download page (instead of the directory listing page),
as it distinguishes between stable and unstable versions. At the
moment, the stable version appears first on the page and the beta
version is second, so I've added a `strategy` block that only uses
the first match. This should be fine for now but it could break if
upstream changes the structure of the page.