pcoipclient 25.03.2
`pcoipclient` is in the autobump list but the latest update, 25.03.2,
is failing because some of the cask URLs are unreachable (whether
temporarily or permanently). While looking into this, I saw that the
`livecheck` block URL redirects to dl.anyware.hp.com, so this updates
the cask `url` and `livecheck` block URL accordingly.

This also updates the `homepage` URL to reference an upstream HP
anywhere page but there may be a better option.