sketch: update livecheck
The existing `livecheck` block for `sketch` checks an appcast file
but the `Sparkle` strategy is incorrectly returning 100.4,180172 as
newest. The 100.4 item has a `pubDate` of 13 May 2025 but the newest
item, 101.9, only has a `pubDate` of 12 May 2025. This resolves the
issue by adding a `strategy` block that simply collects all item
`nice_version` values, ignoring the default strategy sorting.