gitkraken 11.1.1
`gitkraken` is in the autobump list but the 11.1.1 update is failing
because upstream now serves files from api.gitkraken.dev. The file
can be fetched from release.gitkraken.com but the upstream download
page links to api.gitkraken.dev, so this updates the URL accordingly.

Similarly, the `livecheck` block works but it returns 11.1.0 as the
newest version instead of 11.1.1. GitKraken 11.1.1 checks
release.gitkraken.com and that URL returns 11.1.1, as expected, so
this updates the `livecheck` block URL as well.