Standardize regex letter/number character classes
When creating a character class consisting of only letters and
numbers, we typically use `[a-z0-9]` for case-insensitive regexes
(unless working with hex values, where we simply use `\h`). This
updates a couple of instances where the order was different
(`liclipse`) or the capitalization was different (`tap-forms`). [The
`tap-forms` character class ends up matching more characters than
intended (i.e., [, \, ], ^, _, `), so it's more than just a style
issue.]