to-audio-converter: update livecheck, remove autobump
The existing `livecheck` block for `to-audio-converter` is giving an
`Unable to get versions` error because the upstream page that it
checks no longer provides any version information. I wasn't able to
find any other source of version information, so the only alternative
is the `ExtractPlist` strategy. We dislike using this strategy but
it's the only way to identify new version information, as the dmg
filename is unversioned, the app doesn't contain any update logic,
and upstream doesn't publish version information anymore.

This also removes the cask from the autobump list because it uses the
`ExtractPlist` strategy now.