minecraft-server 1.21.6
`minecraft-server` is autobumped but the `livecheck` block is now
returning an `Unable to get versions` error, as the download page HTML
no longer contains any version information or a download link. These
are now fetched in two separate requests using JavaScript, so we have
to do the same. This updates the `livecheck` block to fetch the latest
version and only fetches the download links (to match the hash from
the URL) if the version is newer than the current cask version.

This `strategy` block will return a newer upstream version without the
download URL hash if it fails to fetch the download links JSON, it's
unable to identify the download URL, or unable to match the hash from
the download URL. This would lead to a failed version bump (as the
cask needs the hash to download the `server.jar` file) but this will
at least surface the new version, so we're aware of the issue and can
potentially fix it (e.g., by updating the `livecheck` block).