display-pilot: update livecheck
The existing `livecheck` block for `display-pilot` no longer works
and the only checkable source of version information uses an
asynchronous JSON request that requires a specific referer or it
won't return the content. This sets the `livecheck` block to skip for
now and we can update it with a working check once livecheck has a
`referer` option.