invesalius: add livecheck
`invesalius` is implicitly deprecated because it has a `disable!` call
with a future date, so livecheck skips it as deprecated. This adds a
`livecheck` block to ensure that livecheck will continue checking it
in the mean time. I've used the `GithubLatest` strategy, as some of
the tags with a seemingly stable version format are marked as
"pre-release" on GitHub, so checking the tags isn't sufficient.