openshot-video-editor@daily 3.4.0,14124-6cea273b-0b018e34,release-candidate
The newest daily builds of OpenShot use a file format like
OpenShot-v3.4.0-release-candidate-14124-6cea273b-0b018e34-x86_64.dmg,
so they use a `-release-candidate` suffix instead of `-daily`. This
updates `openshot-video-editor@daily` to the newest daily build,
updating the `url` and `livecheck` block accordingly to handle this
alternative suffix. With this setup, the suffix is only included as
the third `version` part if it's something other than `-daily`.