ripx: update livecheck
The `match.gsub(/<[^>]+?>/, "")` call in the `ripx` `livecheck` block
`strategy` block is flagged as "incomplete multi-character
sanitization" by CodeQL, so it's producing a code scanning alert. This
code isn't intended for sanitization (from a security standpoint) and
is merely a way of allowing us to use a simpler regex (as it's already
pretty involved as it is).

That said, this modifies the related code to repeatedly run until
`#gsub` no longer produces any changes, which should ensure that all
HTML tags are removed and resolve this code scanning alert.