psychopy 2025.2.0
`psychopy` is autobumped but the workflow is failing to update to
2025.2.0 because the filename omits `py` from the `-py3.10` suffix
(upstream is inconsistent from one version to the next).

This also updates the `livecheck` block to check the first-party
download page, as it links to the latest dmg file on GitHub, so we
don't have to use the `GithubReleases` strategy.