bitcoin-core: update livecheck
The existing `livecheck` block for `bitcoin-core` checks the upstream
`/bin/` directory listing page and is returning 29.1 as newest but
this directory doesn't have any expected assets because the version
isn't released yet. This updates the `livecheck` block to return to
checking the first-party download page, which lists the newest
released version.

This `livecheck` block originally checked the download page but it was
modified to check the `/bin/` page instead because the download page
hadn't yet been updated when a contributor created a version bump PR
for 28.1. The page was seemingly updated for the newest version not
too long afterward, so we should have simply waited for it to be
updated as the directory listing page doesn't distinguish between
released and unreleased versions.