sokim: add livecheck
By default, livecheck checks the Git tags for `sokim` and this is
returning 1.2.1 as the newest version but the related release on
GitHub was created four days ago and is marked as pre-release. This
adds a `livecheck` block that uses the `GithubLatest` strategy, as it
seems that upstream creates releases with a stable tag and marks it as
pre-release until the version is released.