fork: update livecheck
The existing `livecheck` block for `fork` uses the `Sparkle` strategy
and this usually works as expected but we're now encountering a
`brew audit` error due to the `sparkle:version` only being `2.56`
instead of `2.56.0`. Since upstream may omit a zero patch from the
`version` value, this updates the `livecheck` block to match the
version from the filename in the `url` and hopefully that will be
more reliable over time.