midi-router-client: add livecheck
By default, livecheck uses the `Sourceforge` strategy for
`midi-router-client` but it's returning 2.4.0 as newest (from the
midi-router-client-2.4.0-Linux.deb file). The only 2.4.0 file at the
moment is the aforementioned deb file and the newest version for macOS
is 2.3.3. This addresses the issue by adding a `livecheck` block with
a regex to restrict matching to only the macOS files.