kdenlive 25.08.1
`kdenlive` is autobumped but the workflow failed to update to 25.08.1
because livecheck is returning an `Unable to get versions` error. The
files for this release include a `_A` suffix at the end of the file
name and the `livecheck` block regex doesn't account for this (as it
diverges from the existing file name format). This updates the cask
to be able to handle a suffix like this (whether it's a new format or
a one-off), so we can update to 25.08.1.