northern-softworks-cache-cleaner 26.0
`northern-softworks-cache-cleaner` is autobumped but livecheck
returns an `Unable to get versions` error because the software name
has been predictably updated to reference the newest macOS version
and the regex no longer matches. This updates the cask accordingly,
which should get it back on track until the next major macOS release.