tencent-docs: deprecate
`tencent-docs` is autobumped but the workflow has been unable to
update to versions beyond 3.10.1 for quite some time. We can
successfully identify new versions but the newer upstream dmg URLs
that I've found are unusable in the cask. I spent an unreasonable
amount of time trying various approaches to try to make this work and
never found a working solution. That's not to say that there isn't a
way but I didn't find it.

With that in mind, this removes the `livecheck` block and deprecates
the cask unless/until there's a way that the cask can use newer
versions.