baidunetdisk 4.54.2
`baidunetdisk` is autobumped but the workflow failed to update to
version 4.54.2 because the Intel dmg file wasn't accessible when the
version was released. The Intel dmg is now available, so this updates
the version accordingly.

Besides that, I checked the response from the `livecheck` block URL
and the URLs for the dmg files have changed, so I've updated the `url`
to align with upstream.