edrawmind 12.4.1
`edrawmind` is autobumped but the `livecheck` block is returning an
innacurate, much lower version because the download page no longer
provides version information. This updates the cask version to the
latest version and modifies the `livecheck` block to check the JSON
data that's used on the support page for downloading products to
populate versions in a select box. This also removes `5378` from the
`version` because it's a fixed product ID that won't change between
versions.