obs-advanced-scene-switcher 1.32.2
`obs-advanced-scene-switcher` is autobumped but the workflow failed
to update to version 1.32.2 because `brew audit` failed due to the
plugin not passing Gatekeeper checks (it's only ad-hoc signed). This
updates the version and adds a `disable!` call with the future date
we've been using in this scenario. Besides that, this also adds a
`depends_on macos:` value (`brew audit` didn't have anything to say
about it but I manually checked the requirement).