propresenter@beta: downgrade monterey version
The version in the `on_monterey` block was carried over from before
there was an `on_ventura` block but the former's `livecheck` block
is returning 18.4,302252046 as newest. This downgrades the version in
the `on_monterey` block to align with the latest version from the
`livecheck` block source. This should keep `brew bump` from mistakenly
updating the `on_monterey` block's version to the same one used in the
`on_ventura` block simply because they're the same string.