brilliant: update livecheck
The existing `livecheck` block for `brilliant` produces an "Unable to
get versions" error, as brilliant.design redirects to
try-brilliant.com but the /api/versions/info path no longer resolves.
The current website only consists of a homepage with a form to join a
waitlist. There doesn't appear to be any available source of
version information at this point, so this updates the `livecheck`
block to skip.