font-brill: disable
`font-brill` is autobumped but the workflow has been failing to check
for new versions because livecheck returns an "Unable to get versions"
error. This has been a persistent issue for some time, as the upstream
website is inaccessible outside of a browser context regardless of
user agent, IP address, etc. Unfortunately, this also applies to the
font zip file and the cask if effectively unusable at this point, so
this disables the cask accordingly.