orka: update livecheck
The existing `livecheck` block for `orka` returns an "Unable to get
versions" error as the downloads page redirects to the main support
page (because the downloads page no longer exists),
support.macstadium.com is behind Cloudflare, and the protections
prevent us from fetching the page. That said, this would fail anyway
because there is no longer a downloads page for Orka CLI versions
before 3.x (for which we have the `orka3` cask). This sets the
`livecheck` block to skip, as Orka CLI 2.x was deprecated when 3.0
was released.