soundsource@test: update livecheck
The existing `livecheck` block for `soundsource@test` produces an
"Unable to get versions" error because the only version in the appcast
is the 5.8.9 stable release, so the regex doesn't match `item.url`.
This adjusts the `strategy` block to return `item.version` if the
regex doesn't match, as this will at least allow us to see and
understand what's happening instead of this appearing as a broken
check.