adobe-dng-converter: update livecheck
The existing `livecheck` block doesn't work because the request needs
to use a "curl/8.7.1" user agent to get the server to return a usable
response but when it does work it returns 17.5 as the newest version
instead of 18.0. This updates the `livecheck` block to check the
redirecting download URL, where we can match the newest version from
the file name in the `Location` header. This approach correctly
returns 18.0 as newest but it also won't work until we can set the
user agent.