tableau-public 2025.2.4
`tableau-public` is autobumped but the workflow failed to update to
version 2025.2.4 because the `livecheck` block is producing an "Unable
to get versions" error, as the website is unreachable unless the
request uses a "curl/8.7.1" user agent. We can't set the user agent in
a `livecheck` block yet (this is forthcoming) but this manually
updates the version in the interim time.