font-amiri: update urls
The GitHub organization for `font-amiri` has changed from alif-type to
aliftype, so the cask `url` is broken and livecheck is returning an
"Unable to get versions" error as a result. This updates the GitHub
URLs, as the homepage contains GitHub URLs using the same
organization. Besides that, this also updates the homepage to resolve
the redirection and align with the website link on the related GitHub
repository.