sonarr 2.18.29
`sonarr` is autobumped but the workflow failed to update to version
2.18.29 because the app fails Gatekeeper checks (due to being
unsigned). This updates the app and adds a `disable!` call with the
future date we've been using for this scenario. This also adds a
`depends_on macos:` value for ARM to resolve the related `brew audit`
error. The Intel app has a macOS requirement of 10.14 or later, so it
can use the implicit macOS requirement but we have to add an explicit
dependency to pass `brew style` (as we can't have only an ARM value).