replicator 8.4.1
`replicator` is autobumped but the workflow failed to update to
version 8.4.1 because `brew audit` failed with a "Cask is deprecated
because it failed Gatekeeper checks but all artifacts now pass!"
error. I've manually confirmed that the 8.4.1 app passes Gatekeeper
checks, so this updates the version and removes the `disable!` call.