canon-eos-utility 3.20.10.1
`canon-eos-utility` is autobumped but the workflow failed to update to
version 3.20.10.1 because the `livecheck` block is producing an
"Unable to get versions" error, as the file name format has changed
and the regex does not match it. This updates the version, adjusts the
`url`, and tweaks the `livecheck` block regex accordingly.