dvdstyler: update livecheck
The existing `livecheck` block for `dvdstyler` checks the first-party
download page but this no longer contains a link to the dmg file
(only a generic link to a dAppCDN page). The cask uses a file from
SourceForge and the download page links to the SourceForge project as
the source for all releases, so this updates the `livecheck` block to
use the `Sourceforge` strategy to check for new versions, restricting
the check to the stable `dvdstyler` directory (omitting the sibling
`dvdstyler-devel` and `OldFiles` directories).