mobirise: update livecheck
The existing `livecheck` block for `mobirise` is returning an "Unable
to get versions" error because the server returns a 404 (Not Found)
response. This URL had been redirecting to history.mobirisesite.com
for a few months but it seems that this redirection has been removed.
This addresses the issue by updating the `livecheck` block URL.