amazon-q 1.19.7
`amazon-q` is autobumped but the workflow tried to update to a 1.20.0
version instead of 1.19.7. The JSON data that the `livecheck` block
is checking contains versions for other software (i.e., 1.20.0 is a
kirocli version), so we have to specifically target Q versions. This
updates the cask and introduces some logic in the `livecheck` block's
`strategy` block to only surface versions providing a Q dmg package.
This could silently break in the future if the file name format
changes (e.g., adding a suffix, different extension, etc.) but this
works for now.